+++ packages/backend/src/lib/utils.ts
@@ -21,16 +21,16 @@ import {
     commonTerms,
     getCircuitsCollectionPath,
     getContributionsCollectionPath,
-    getTimeoutsCollectionPath
-} from "@p0tion/actions/src"
+    getTimeoutsCollectionPath,
+    CeremonyState,
+    finalContributionIndex,
+    CircuitDocument
+} from "@p0tion/actions"
 import fetch from "@adobe/node-fetch-retry"
-import { CeremonyState } from "@p0tion/actions/src/types/enums"
 import path from "path"
 import os from "os"
-import { finalContributionIndex } from "@p0tion/actions/src/helpers/constants"
 import { COMMON_ERRORS, logAndThrowError, SPECIFIC_ERRORS } from "./errors"
 import { getS3Client } from "./services"
-import { CircuitDocument } from "@p0tion/actions/src/types"
 
 dotenv.config()
 
@@ -85,7 +85,9 @@ export const getCeremonyCircuits = async (ceremonyId: string): Promise<Array<Que
 
     if (!querySnap.docs) logAndThrowError(SPECIFIC_ERRORS.SE_CONTRIBUTE_NO_CEREMONY_CIRCUITS)
 
-    return querySnap.docs
+    return querySnap.docs.sort(
+        (a: DocumentData, b: DocumentData) => a.data().sequencePosition - b.data().sequencePosition
+    )
 }
 
 /**
@@ -198,6 +200,7 @@ export const downloadArtifactFromS3Bucket = async (bucketName: string, objectKey
     const url = await getSignedUrl(client, command, { expiresIn: Number(process.env.AWS_PRESIGNED_URL_EXPIRATION) })
 
     // Execute download request.
+    // @ts-ignore
     const response: any = await fetch(url, {
         method: "GET",
         headers: {
@@ -219,7 +222,12 @@ export const downloadArtifactFromS3Bucket = async (bucketName: string, objectKey
  * @param objectKey <string> - the unique key to identify the object inside the given AWS S3 bucket.
  * @param localFilePath <string> - the local path where the file to be uploaded is stored.
  */
-export const uploadFileToBucket = async (bucketName: string, objectKey: string, localFilePath: string) => {
+export const uploadFileToBucket = async (
+    bucketName: string,
+    objectKey: string,
+    localFilePath: string,
+    isPublic: boolean = false
+) => {
     // Prepare AWS S3 client instance.
     const client = await getS3Client()
 
@@ -227,12 +235,18 @@ export const uploadFileToBucket = async (bucketName: string, objectKey: string, 
     const contentType = mime.lookup(localFilePath) || ""
 
     // Prepare command.
-    const command = new PutObjectCommand({ Bucket: bucketName, Key: objectKey, ContentType: contentType })
+    const command = new PutObjectCommand({
+        Bucket: bucketName,
+        Key: objectKey,
+        ContentType: contentType,
+        ACL: isPublic ? "public-read" : "private"
+    })
 
     // Generate a pre-signed url for uploading the file.
     const url = await getSignedUrl(client, command, { expiresIn: Number(process.env.AWS_PRESIGNED_URL_EXPIRATION) })
 
     // Execute upload request.
+    // @ts-ignore
     const response = await fetch(url, {
         method: "PUT",
         body: readFileSync(localFilePath),
@@ -318,27 +332,24 @@ export const getFinalContribution = async (
  * @param circuitDocument <CircuitDocument> - the circuit document to be encoded.
  * @returns <CircuitDocument> - the circuit document encoded.
  */
-export const htmlEncodeCircuitData = (
-    circuitDocument: CircuitDocument
-): CircuitDocument => {
-    return {
-        ...circuitDocument,
-        description: encode(circuitDocument.description),
-        name: encode(circuitDocument.name),
-        prefix: encode(circuitDocument.prefix)
-    }
-}
+export const htmlEncodeCircuitData = (circuitDocument: CircuitDocument): CircuitDocument => ({
+    ...circuitDocument,
+    description: encode(circuitDocument.description),
+    name: encode(circuitDocument.name),
+    prefix: encode(circuitDocument.prefix)
+})
 
 /**
  * Fetch the variables related to GitHub anti-sybil checks
  * @returns <any> - the GitHub variables.
  */
-export const getGitHubVariables = () : any => {
+export const getGitHubVariables = (): any => {
     if (
-        !process.env.GITHUB_MINIMUM_FOLLOWERS || 
+        !process.env.GITHUB_MINIMUM_FOLLOWERS ||
         !process.env.GITHUB_MINIMUM_FOLLOWING ||
         !process.env.GITHUB_MINIMUM_PUBLIC_REPOS
-    ) logAndThrowError(COMMON_ERRORS.CM_WRONG_CONFIGURATION)
+    )
+        logAndThrowError(COMMON_ERRORS.CM_WRONG_CONFIGURATION)
 
     return {
         minimumFollowers: Number(process.env.GITHUB_MINIMUM_FOLLOWERS),