+++ packages/phase2cli/CHANGELOG.md
@@ -3,40 +3,98 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.2.0](https://github.com/quadratic-funding/mpc-phase2-suite/cli/compare/v0.1.0...v0.2.0) (2023-04-04)
-
-
-### ⚠ BREAKING CHANGES
-
-* constants are now part of actions package
-
-### Features
-
-* add public finalization beacon value and hash for final contribution ([74dfc07](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/74dfc074c745c385787c15d84dbfb5a3a5f20cf8))
-* add wasm support at setup time ([53eeddd](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/53eeddd14e18504ae81cae57c0ee846d4b9935ad))
-* enable creation of public attestation gist even after contribution ([54db59c](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/54db59cc7a07dd291e37df2b36f5c5b1a457eed2))
-* **verification:** implemented actions functions on phase2cli ([7ae6da3](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/7ae6da37201bb644bababa45adf40890173f2c25))
+## [0.4.2](https://github.com/privacy-scaling-explorations/p0tion/compare/v0.3.0...v0.4.2) (2023-05-31)
 
 
 ### Bug Fixes
 
-* **ceremony verification:** add missing d.ts files ([7cec92e](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/7cec92e99ded7a9295f3ba1020d16f0bb611eb58))
-* **ceremony verification:** remove redundant argument from export verifier function ([ec351de](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/ec351def2a1104b825ead8619d65215431b38abb))
-* error when building in CI; review suggestions ([c811c43](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/c811c43fa01f697b97856ae6cc80e8e87565139b))
-* make spinner to stop properly ([d47de71](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/d47de71fa63c3e4ffabdda4155d83debf2e4947f))
-* missing pre-condition when resuming during computing step; wrong pre-condition on related cf ([90c8729](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/90c87295b771d099cffc43880e4e962cd64fd330))
-* small bug fixes and minor changes ([0a0d44a](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/0a0d44aa4f74aea2140ee68bc5547009611f2372))
-* **verification:** amend after PR review ([35ecf47](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/35ecf47d90a1f173b1cbcfa338a8b528899633b2))
-* **verification:** fix wrong path in function call ([cc9a991](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/cc9a9916aeeca5de12a09bb053b213e3111522df))
-* wrong Date type for start/end ceremony dates expressed in ms ([655a02e](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/655a02ed33f8e36f9224a1c299320f9e50504955))
-* wrong version for peer dependency actions package ([4975d96](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/4975d96928ec2f02dab99c493f9b4e65ff6ec983))
+* duplicate messages when waiting for contribution; remove wrong listener ([61fbd19](https://github.com/privacy-scaling-explorations/p0tion/commit/61fbd19eb13f35d34963fbd350441f54cd8e1c91))
+* spinner does not stop correctly after contribution verification ([345bdcb](https://github.com/privacy-scaling-explorations/p0tion/commit/345bdcb8a09915aba29b37ac988b7109a67490e0))
+* unconsistency when resuming a contribution ([07fc79b](https://github.com/privacy-scaling-explorations/p0tion/commit/07fc79b8415935eb7f5d0f11372f60d04f7d997c))
 
+
+
+## [0.4.1](https://github.com/privacy-scaling-explorations/p0tion/compare/v0.3.0...v0.4.1) (2023-05-31)
+
+
+### Bug Fixes
+
+* duplicate messages when waiting for contribution; remove wrong listener ([61fbd19](https://github.com/privacy-scaling-explorations/p0tion/commit/61fbd19eb13f35d34963fbd350441f54cd8e1c91))
+* spinner does not stop correctly after contribution verification ([345bdcb](https://github.com/privacy-scaling-explorations/p0tion/commit/345bdcb8a09915aba29b37ac988b7109a67490e0))
+* unconsistency when resuming a contribution ([07fc79b](https://github.com/privacy-scaling-explorations/p0tion/commit/07fc79b8415935eb7f5d0f11372f60d04f7d997c))
+
+
+
+## [0.4.0](https://github.com/privacy-scaling-explorations/p0tion/compare/v0.3.0...v0.4.0) (2023-05-19)
+
+### Bug Fixes
+
+-   duplicate messages when waiting for contribution; remove wrong listener ([61fbd19](https://github.com/privacy-scaling-explorations/p0tion/commit/61fbd19eb13f35d34963fbd350441f54cd8e1c91))
+-   spinner does not stop correctly after contribution verification ([345bdcb](https://github.com/privacy-scaling-explorations/p0tion/commit/345bdcb8a09915aba29b37ac988b7109a67490e0))
+-   unconsistency when resuming a contribution ([07fc79b](https://github.com/privacy-scaling-explorations/p0tion/commit/07fc79b8415935eb7f5d0f11372f60d04f7d997c))
+
+## [0.3.5](https://github.com/privacy-scaling-explorations/p0tion/compare/v0.3.0...v0.3.5) (2023-05-17)
+
+### Bug Fixes
+
+-   duplicate messages when waiting for contribution; remove wrong listener ([61fbd19](https://github.com/privacy-scaling-explorations/p0tion/commit/61fbd19eb13f35d34963fbd350441f54cd8e1c91))
+-   spinner does not stop correctly after contribution verification ([345bdcb](https://github.com/privacy-scaling-explorations/p0tion/commit/345bdcb8a09915aba29b37ac988b7109a67490e0))
+
+## [0.3.2](https://github.com/privacy-scaling-explorations/p0tion/compare/v0.3.0...v0.3.2) (2023-05-17)
+
+### Bug Fixes
+
+-   duplicate messages when waiting for contribution; remove wrong listener ([61fbd19](https://github.com/privacy-scaling-explorations/p0tion/commit/61fbd19eb13f35d34963fbd350441f54cd8e1c91))
+-   spinner does not stop correctly after contribution verification ([345bdcb](https://github.com/privacy-scaling-explorations/p0tion/commit/345bdcb8a09915aba29b37ac988b7109a67490e0))
+
+## [0.3.1](https://github.com/privacy-scaling-explorations/p0tion/compare/v0.3.0...v0.3.1) (2023-05-16)
+
+### Bug Fixes
+
+-   spinner does not stop correctly after contribution verification ([345bdcb](https://github.com/privacy-scaling-explorations/p0tion/commit/345bdcb8a09915aba29b37ac988b7109a67490e0))
+
+## [0.3.0](https://github.com/privacy-scaling-explorations/p0tion/compare/v0.2.0...v0.3.0) (2023-05-15)
+
+### Features
+
+-   add typedoc generation ([6fee9d4](https://github.com/privacy-scaling-explorations/p0tion/commit/6fee9d422f4331997ebdbc152ed0b3fd36f43ede))
+
+### Bug Fixes
+
+-   bump 0.3.0; minor missing imports for enums ([79faae9](https://github.com/privacy-scaling-explorations/p0tion/commit/79faae92a04f4b6976645057623cf8f951116eb2))
+-   missing skipLibCheck option was causing docs script to exit ([7433e38](https://github.com/privacy-scaling-explorations/p0tion/commit/7433e38395382f9f49e7eb499cbfb114bd6efd58))
+-   set cli as a es module ([37f4351](https://github.com/privacy-scaling-explorations/p0tion/commit/37f43518169194b39479cfc194ec8e29f3f88b3c))
+-   wrong path for environment config file ([75bbd98](https://github.com/privacy-scaling-explorations/p0tion/commit/75bbd98b129754bbd93b0160afcdd1abd4e21f49))
+
+## [0.2.0](https://github.com/quadratic-funding/mpc-phase2-suite/cli/compare/v0.1.0...v0.2.0) (2023-04-04)
+
+### ⚠ BREAKING CHANGES
+
+-   constants are now part of actions package
+
+### Features
+
+-   add public finalization beacon value and hash for final contribution ([74dfc07](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/74dfc074c745c385787c15d84dbfb5a3a5f20cf8))
+-   add wasm support at setup time ([53eeddd](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/53eeddd14e18504ae81cae57c0ee846d4b9935ad))
+-   enable creation of public attestation gist even after contribution ([54db59c](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/54db59cc7a07dd291e37df2b36f5c5b1a457eed2))
+-   **verification:** implemented actions functions on phase2cli ([7ae6da3](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/7ae6da37201bb644bababa45adf40890173f2c25))
+
+### Bug Fixes
+
+-   **ceremony verification:** add missing d.ts files ([7cec92e](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/7cec92e99ded7a9295f3ba1020d16f0bb611eb58))
+-   **ceremony verification:** remove redundant argument from export verifier function ([ec351de](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/ec351def2a1104b825ead8619d65215431b38abb))
+-   error when building in CI; review suggestions ([c811c43](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/c811c43fa01f697b97856ae6cc80e8e87565139b))
+-   make spinner to stop properly ([d47de71](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/d47de71fa63c3e4ffabdda4155d83debf2e4947f))
+-   missing pre-condition when resuming during computing step; wrong pre-condition on related cf ([90c8729](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/90c87295b771d099cffc43880e4e962cd64fd330))
+-   small bug fixes and minor changes ([0a0d44a](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/0a0d44aa4f74aea2140ee68bc5547009611f2372))
+-   **verification:** amend after PR review ([35ecf47](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/35ecf47d90a1f173b1cbcfa338a8b528899633b2))
+-   **verification:** fix wrong path in function call ([cc9a991](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/cc9a9916aeeca5de12a09bb053b213e3111522df))
+-   wrong Date type for start/end ceremony dates expressed in ms ([655a02e](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/655a02ed33f8e36f9224a1c299320f9e50504955))
+-   wrong version for peer dependency actions package ([4975d96](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/4975d96928ec2f02dab99c493f9b4e65ff6ec983))
 
 ### Code Refactoring
 
-* update and move constants to actions package ([e1a98a8](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/e1a98a8d4b33c589a4a32300e9ad03c9a647c05b))
-
-
+-   update and move constants to actions package ([e1a98a8](https://github.com/quadratic-funding/mpc-phase2-suite/cli/commit/e1a98a8d4b33c589a4a32300e9ad03c9a647c05b))
 
 ## [0.1.0](https://github.com/quadratic-funding/mpc-phase2-suite/cli/compare/v0.0.1...v0.1.0) (2023-01-24)